remove code duplication between Scala specific versions to satisfy SonarCloud
